requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
colorama==0.4.6
contextlib2==21.6.0
cryptography==44.0.1
Django==5.1.6
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-ninja==1.3.0
django-ninja-extra==0.22.3
django-ninja-jwt==5.3.5
django-phonenumber-field==8.0.0
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
injector==0.22.0
kombu==5.4.2
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==24.2
phonenumbers==8.13.55
pillow==11.1.0
prompt_toolkit==3.0.50
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
